Print --list-network-interfaces then exit (without launching server)
