+++ lib/cli-app.js
@@ -170,6 +170,8 @@ class LwsCli {
       this.showConfig(config)
     } else if (config.version) {
       this.showVersion()
+    } else if (config.listNetworkInterfaces) {
+      this.printNetworkInterfaces(config.hostname)
     } else {
       return this.launchServer(config)
     }
@@ -216,6 +218,14 @@ class LwsCli {
   showUsage () {
     const commandLineUsage = require('command-line-usage')
     this.log(commandLineUsage(this.usage()))
+  }
+
+  printNetworkInterfaces (hostname) {
+    const ansi = require('ansi-escape-sequences')
+    const util = require('./util')
+    const ipList = util.getIPList(hostname)
+    const interfaces = ipList.map(iface => `- ${ansi.format(iface.name, 'bold')}: ${iface.address}`).join('\n')
+    this.log(`\n${ansi.format('Available network interfaces', 'underline')}\n${interfaces}\n`)
   }
 }
