+++ lib/cli-args.js
@@ -117,5 +117,10 @@ module.exports = [
     type: Boolean,
     description: 'Automatically open the default system browser.',
     section: 'core'
+  },
+  {
+    name: 'list-network-interfaces',
+    type: Boolean,
+    description: 'Print a list of available network interfaces.'
   }
 ]
