+++ lib/view/cli-view.js
@@ -15,11 +15,6 @@ class CliView {
   optionDefinitions () {
     return [
       {
-        name: 'list-network-interfaces',
-        type: Boolean,
-        description: 'Outputs a list of host\'s network interfaces.'
-      },
-      {
         name: 'qr',
         type: String,
         description: 'Outputs a QRCode with the network_interface listener address. If the network_interface arg is not provided, it will be defaulted to the first private address (IETF RFC1918) found among the network interfaces.',
@@ -50,13 +45,7 @@ class CliView {
     ipList = ipList
       .map(iface => ansi.format(iface.url, 'underline'))
       .join(', ')
-    _log.get(this)(`\nListening on ${ipList}\n`)
-  }
-
-  printNetworkInterfaces (ipList) {
-    const ansi = require('ansi-escape-sequences')
-    const interfaces = ipList.map(iface => `- ${ansi.format(iface.name, 'bold')}: ${iface.address}`).join('\n')
-    _log.get(this)(`\n${ansi.format('Available network interfaces', 'underline')}\n${interfaces}\n`)
+    _log.get(this)(`Listening on ${ipList}`)
   }
 
   resolvePrivateAddress (ipList) {
@@ -140,9 +129,6 @@ class CliView {
       const { printError } = require('../util')
       printError(value, 'Middleware error', _logError.get(this))
     } else if (key === 'server.listening') {
-      if (config.listNetworkInterfaces) {
-        this.printNetworkInterfaces(value)
-      }
       if (config.qr === null || config.qr) {
         this.printAddressQRCode(config.qr, value)
       }
